requirements.txt
Django==1.8
django-crispy-forms==1.5.2
django-registration-redux==1.2
django-tinymce==2.2.0
Pillow==2.8.1
psycopg2==2.6.1
requests==2.7.0
virtualenv==13.1.2
